+++ src/main/java/com/apps/pochak/comment/dto/response/CommentElements.java
@@ -1,7 +1,7 @@
 package com.apps.pochak.comment.dto.response;
 
 import com.apps.pochak.comment.domain.Comment;
-import com.apps.pochak.global.PageInfo;
+import com.apps.pochak.global.util.PageInfo;
 import com.apps.pochak.member.domain.Member;
 import lombok.AllArgsConstructor;
 import lombok.Data;
